mcp-server/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "nutrition-tools"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[project.scripts]
nutrition-tools = "nutrition_tools.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
